Move Site Tour to it's own component
